refactor(website): introduce view keys - don't search on which page you are
The old `Routing` class used to have a collection of all "views".
Usually, when one wanted to do something with routing, one had to search all views.

This shifts the perspective: You know on which page you are (because Astro already did the routing).
So you also know which view to use. The view keys are a text representation so they can be handed between
client- and server-side code easily (the views can't because they carry function).
"routing" now essentially became a factory that constructs the view for the corresponding key.