feat(website): add 'Sub-lineages Neststrain Clade' for SARS-CoV-2 (#559)
The reason to implement it using a separate component instantiation
rather than adding another `LineageFilterConfig` to
`CovidConstants.lineageFilters` (which adds another column to the
existing output) is that the ~corresponding Nextstrain Clade for a
Nextclade pango lineage is not actually associated with the same data,
thus it would be wrong. Also, the bar chart currently only shows the
first column, thus would not show the Nextstrain Clade names.

Also changing the title of the existing view. (The same value is
already in the `LineageFilterConfig` as `placeholderText`, but the
`lineageFilters`/`nextstrainCladeLineageFilters` fields have an array
of those, thus can't extract it reliably.)